Language.PlutusTx.Prelude

Documentation

The PlutusTx Prelude is a collection of useful functions that work with builtin Haskell data types such as Maybe and [] (list).

Functions from the Prelude can be used with the the typed Template Haskell splice operator $$():

  import qualified Language.PlutusTx.Prelude as P

  [||  $$(P.traceH) "plutus" ... ||]

toPlutusString :: Q (TExp (String -> String))

Convert a Haskell String into a PlutusTx String.

trace :: Q (TExp (String -> a -> a))

Emit the given string as a trace message before evaluating the argument.

traceH :: Q (TExp (String -> a -> a))

A version of trace that takes a Haskell String.

Error

error :: Q (TExp (() -> a))

Terminate the evaluation of the script with an error message

Boolean operators

and :: Q (TExp (Bool -> Bool -> Bool))

Logical AND

>>> $$([|| $$(and) True False ||])
False

or :: Q (TExp (Bool -> Bool -> Bool))

Logical OR

>>> $$([|| $$(or) True False ||])
True

not :: Q (TExp (Bool -> Bool))

Logical negation

>>> $$([|| $$(not) True ||])
False

Numbers

min :: Q (TExp (Int -> Int -> Int))

The smaller of two Ints

>>> $$([|| $$(min) 10 5 ||])
5

max :: Q (TExp (Int -> Int -> Int))

The larger of two Ints

>>> $$([|| $$(max) 10 5 ||])
10

Maybe

isJust :: Q (TExp (Maybe a -> Bool))

Check if a Maybe a is Just a

>>> $$([|| $$(isJust) Nothing ||])
False
>>> $$([|| $$(isJust) (Just "plutus") ||])
True

isNothing :: Q (TExp (Maybe a -> Bool))

Check if a Maybe a is Nothing

>>> $$([|| $$(isNothing) Nothing ||])
True
>>> $$([|| $$(isNothing) (Just "plutus") ||])
False

maybe :: Q (TExp (b -> (a -> b) -> Maybe a -> b))

PlutusTx version of maybe.

>>> $$([|| $$(maybe) "platypus" (\s -> s) (Just "plutus") ||])
"plutus"
>>> $$([|| $$(maybe) "platypus" (\s -> s) Nothing ||])
"platypus"

Lists

map :: Q (TExp ((a -> b) -> [a] -> [b]))

PlutusTx version of map.

>>> $$([|| $$(map) (\i -> i + 1) [1, 2, 3] ||])
[2,3,4]

foldr :: Q (TExp ((a -> b -> b) -> b -> [a] -> b))

PlutusTx version of foldr.

>>> $$([|| $$(foldr) (\i s -> s + i) 0 [1, 2, 3, 4] ||])
10

foldl :: Q (TExp ((b -> a -> b) -> b -> [a] -> b))

PlutusTx version of foldl.

>>> $$([|| $$(foldl) (\s i -> s + i) 0 [1, 2, 3, 4] ||])
10

length :: Q (TExp ([a] -> Int))

length xs is the number of elements in xs.

>>> $$([|| $$(length) [1, 2, 3, 4] ||])
4

all :: Q (TExp ((a -> Bool) -> [a] -> Bool))

PlutusTx version of all.

>>> $$([|| $$(all) (\i -> i > 5) [6, 8, 12] ||])
True

Hashes

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

txhash :: Q (TExp ByteString)

sha2_256 :: Q (TExp (ByteString -> ByteString))

The double SHA256 hash of a ByteString

sha3_256 :: Q (TExp (ByteString -> ByteString))

The triple SHA256 hash of a ByteString

equalsByteString :: Q (TExp (ByteString -> ByteString -> Bool))

Check if two ByteStrings are equal

takeByteString :: Q (TExp (Int -> ByteString -> ByteString))

Returns the n length prefix of a ByteString

dropByteString :: Q (TExp (Int -> ByteString -> ByteString))

Returns the suffix of a ByteString after n elements

concatenate :: Q (TExp (ByteString -> ByteString -> ByteString))

Concatenates two ByteStrings together.