Language.PlutusTx.TH

Documentation

compile :: Q (TExp a) -> Q (TExp (CompiledCode a))

Compile a quoted Haskell expression into a corresponding Plutus Core program.

compileUntyped :: Q Exp -> Q Exp

Compile a quoted Haskell expression into a corresponding Plutus Core program.

data CompiledCode a

A compiled Plutus Tx program. The type parameter inicates the type of the Haskell expression that was compiled, and hence the type of the compiled code.

getSerializedPlc :: CompiledCode a -> ByteString

getSerializedPir :: CompiledCode a -> ByteString

getPlc :: CompiledCode a -> Program TyName Name ()

getPir :: CompiledCode a -> Program TyName Name ()