PlutusTx.Blueprint.Contract

Documentation

data ContractBlueprint where

A blueprint of a smart contract, as defined by the CIP-0057

The referencedTypes type variable is used to track the types used in the contract making sure their schemas are included in the blueprint and that they are referenced in a type-safe way. See Note [Unrolling types] for more details.

Constructors

MkContractBlueprint

Fields

:: forall referencedTypes. { contractId :: Maybe Text

An optional identifier for the contract.

, contractPreamble :: Preamble

An object with meta-information about the contract.

, contractValidators :: ToJSON ContractBlueprint

Instance details

Defined in PlutusTx.Blueprint.Contract

Methods

toJSON :: ContractBlueprint -> Value

toEncoding :: ContractBlueprint -> Encoding

toJSONList :: [ContractBlueprint] -> Value

toEncodingList :: [ContractBlueprint] -> Encoding

omitField :: ContractBlueprint -> Bool