PlutusTx.Blueprint.Definition

Description

This module provides a functionality to derive and reference schema definitions.

Documentation

data DefinitionId

A reference to a Schema definition.

class AsDefinitionId a where

Minimal complete definition

Nothing

Methods

definitionId :: DefinitionId

default definitionId :: Typeable a => DefinitionId

data Definitions (ts :: [Type]) where

A registry of schema definitions.

Constructors

NoDefinitions :: Definitions '[]

AddDefinition :: Definition t ts -> Definitions ts -> Definitions (t ': ts)

data Definition t ts

A schema definition of a type t with a list of referenced types ts.

Constructors

MkDefinition DefinitionId (Schema ts)

definition :: forall t ts. (AsDefinitionId t, HasSchema t ts) => Definition t ts

Construct a schema definition.

definitionRef :: forall t ts. (AsDefinitionId t, HasSchemaDefinition t ts) => Schema ts

Construct a schema that is a reference to a schema definition.

addDefinition :: Definitions ts -> Definition t ts -> Definitions (t ': ts)

Add a schema definition to a registry.

definitionsToMap :: Definitions ts -> (forall xs. Schema xs -> v) -> Map DefinitionId v

type family HasSchemaDefinition n xs where ...

A constraint that checks if a schema definition is present in a list of schema definitions. Gives a user-friendly error message if the schema definition is not found.

Equations

HasSchemaDefinition x (x ': xs) = ()

HasSchemaDefinition x (_ ': xs) = HasSchemaDefinition x xs

HasSchemaDefinition n xs = TypeError (ShowType n :<>: Text " type was not found in the list of types having schema definitions.")

Type-level utilities

type family Unroll (p :: Type) :: [Type] where ...

Unroll a type into a list of all nested types (including the type itself).

If a type doesn't have a generic representation, then this type family gets "stuck". The good news is that for the purpose of deriving schema definitions, we only need to consider types that are either end-user defined (and therefore have a generic representation) or built-in types that we explicitly list here as terminals in order not to get "stuck".

Equations

Unroll Int = '[Int]

Unroll Integer = '[Integer]

Unroll Text = '[Text]

Unroll BuiltinData = '[BuiltinData]

Unroll BuiltinUnit = '[BuiltinUnit]

Unroll BuiltinString = '[BuiltinString]

Unroll (BuiltinList a) = Prepend (BuiltinList a) (GUnroll (Rep a))

Unroll BuiltinByteString = '[BuiltinByteString]

Unroll p = Prepend p (GUnroll (Break (NoGeneric p) (Rep p)))

type family UnrollAll xs :: [Type] where ...

Equations

UnrollAll '[] = '[]

UnrollAll (x ': xs) = Concat (Unroll x) (UnrollAll xs)

class Unrollable ts where

This class and its two instances are used internally to derive Definitions for a given list of types.

Methods

unroll :: Definitions ts

deriveDefinitions :: forall ts. Unrollable (UnrollAll ts) => Definitions (UnrollAll ts)