PlutusTx.Blueprint.Parameter

Documentation

data ParameterBlueprint (referencedTypes :: [Type])

Blueprint that defines validator's compile-time parameter.

The referencedTypes phantom type parameter is used to track the types used in the contract making sure their schemas are included in the blueprint and that they are referenced in a type-safe way.

Constructors

MkParameterBlueprint

Fields

parameterTitle :: Maybe Text

An informative description of the parameter.

parameterPurpose :: ToJSON a => Set a -> Maybe Value