PlutusTx.BuiltinList

Description

Functions operating on BuiltinList.

Documentation

data BuiltinList a

type Unroll (BuiltinList a)

Instance details

Defined in PlutusTx.Blueprint.Definition.Unroll

type Unroll (BuiltinList a) = Insert (BuiltinList a) (Unrolled a)

type FromBuiltin (BuiltinList a)

Instance details

Defined in PlutusTx.Builtins.HasBuiltin

type FromBuiltin (BuiltinList a) = [FromBuiltin a]

cons :: forall a. a -> BuiltinList a -> BuiltinList a

Plutus Tx version of : for BuiltinList.

uncons :: forall a. BuiltinList a -> Maybe (a, BuiltinList a)

Plutus Tx version of uncons for BuiltinList.

empty :: forall a. MkNil a => BuiltinList a

Plutus Tx version of '[]' for BuiltinList.

singleton :: forall a. MkNil a => a -> BuiltinList a

Make a list with one element.

null :: forall a. BuiltinList a -> Bool

Plutus Tx version of null for BuiltinList.

caseList' :: forall a r. r -> (a -> BuiltinList a -> r) -> BuiltinList a -> r

caseList :: forall a r. (() -> r) -> (a -> BuiltinList a -> r) -> BuiltinList a -> r

map :: forall a b. MkNil b => (a -> b) -> BuiltinList a -> BuiltinList b

Plutus Tx version of map for BuiltinList.

elem :: forall a. Eq a => a -> BuiltinList a -> Bool

Does the element occur in the list?

find :: forall a. (a -> Bool) -> BuiltinList a -> Maybe a

Returns the leftmost element matching the predicate, or Nothing if there's no such element.

any :: forall a. (a -> Bool) -> BuiltinList a -> Bool

Determines whether any element of the structure satisfies the predicate.

all :: forall a. (a -> Bool) -> BuiltinList a -> Bool

Determines whether all elements of the list satisfy the predicate.

(!!) :: forall a. BuiltinList a -> Integer -> a infixl 9

Get the element at a given index. This function throws an error if the index is negative or larger than the length of the list.

(++) :: forall a. BuiltinList a -> BuiltinList a -> BuiltinList a infixr 5

Plutus Tx version of (++) for BuiltinList.

(<|) :: forall a. a -> BuiltinList a -> BuiltinList a infixr 5

Infix version of cons.

append :: forall a. BuiltinList a -> BuiltinList a -> BuiltinList a

Plutus Tx version of append for BuiltinList.

findIndices :: forall a. (a -> Bool) -> BuiltinList a -> BuiltinList Integer

Plutus Tx version of findIndices for BuiltinList.

filter :: forall a. MkNil a => (a -> Bool) -> BuiltinList a -> BuiltinList a

Plutus Tx version of filter for BuiltinList.

mapMaybe :: forall a b. MkNil b => (a -> Maybe b) -> BuiltinList a -> BuiltinList b

Plutus Tx version of mapMaybe for BuiltinList.

length :: forall a. BuiltinList a -> Integer

Plutus Tx version of length for BuiltinList.

and :: BuiltinList Bool -> Bool

Returns the conjunction of a list of Bools.

or :: BuiltinList Bool -> Bool

Returns the disjunction of a list of Bools.

notElem :: forall a. Eq a => a -> BuiltinList a -> Bool

The negation of elem.

foldr :: forall a b. (a -> b -> b) -> b -> BuiltinList a -> b

Plutus Tx version of foldr for BuiltinList.

foldl :: forall a b. (b -> a -> b) -> b -> BuiltinList a -> b

Plutus Tx velsion of foldl for BuiltinList.

concat :: forall a. MkNil a => BuiltinList (BuiltinList a) -> BuiltinList a

Plutus Tx version of concat for BuiltinList.

concatMap :: forall a b. MkNil b => (a -> BuiltinList b) -> BuiltinList a -> BuiltinList b

Plutus Tx version of concatMap for BuiltinList.

listToMaybe :: forall a. BuiltinList a -> Maybe a

Plutus Tx version of listToMaybe for BuiltinList.

uniqueElement :: forall a. BuiltinList a -> Maybe a

Return the element in the list, if there is precisely one.

revAppend :: forall a. BuiltinList a -> BuiltinList a -> BuiltinList a

Cons each element of the first list to the second one in reverse order (i.e. the last element of the first list is the head of the result).

revAppend xs ys === reverse xs ++ ys 

reverse :: forall a. MkNil a => BuiltinList a -> BuiltinList a

Plutus Tx version of reverse for BuiltinList.

replicate :: forall a. MkNil a => Integer -> a -> BuiltinList a

Plutus Tx version of replicate for BuiltinList.

findIndex :: forall a. (a -> Bool) -> BuiltinList a -> Maybe Integer

Plutus Tx version of findIndex.

head :: forall a. BuiltinList a -> a

Plutus Tx version of head for BuiltinList.

last :: forall a. BuiltinList a -> a

Plutus Tx version of last for BuiltinList.

tail :: forall a. BuiltinList a -> BuiltinList a

Plutus Tx version of tail for BuiltinList.

take :: forall a. MkNil a => Integer -> BuiltinList a -> BuiltinList a

Plutus Tx version of take for BuiltinList.

drop :: forall a. MkNil a => Integer -> BuiltinList a -> BuiltinList a

Plutus Tx version of drop for BuiltinList.

dropWhile :: forall a. (a -> Bool) -> BuiltinList a -> BuiltinList a

Plutus Tx version of dropWhile for BuiltinList.

elemBy :: forall a. (a -> a -> Bool) -> a -> BuiltinList a -> Bool

Plutus Tx version of elemBy for BuiltinList.

nub :: forall a. (Eq a, MkNil a) => BuiltinList a -> BuiltinList a

Plutus Tx version of nub for BuiltinList.

nubBy :: forall a. MkNil a => (a -> a -> Bool) -> BuiltinList a -> BuiltinList a

Plutus Tx version of nubBy for BuiltinList.

zipWith :: forall a b c. MkNil c => (a -> b -> c) -> BuiltinList a -> BuiltinList b -> BuiltinList c

Plutus Tx version of zipWith for BuiltinList.