PlutusTx.Coverage

Decoder s [CoverageAnnotation]

type Rep CoverageAnnotation

Instance details

Defined in PlutusTx.Coverage

type Rep CoverageAnnotation = D1 ('MetaData "CoverageAnnotation" "PlutusTx.Coverage" "plutus-tx-1.60.0.0-inplace" 'False) (C1 ('MetaCons "CoverLocation" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 CovLoc)) :+: C1 ('MetaCons "CoverBool" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 CovLoc) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 Bool)))

newtype CoverageIndex

This type keeps track of all coverage annotations and where they have been inserted / what annotations are expected to be found when executing a piece of code.

Constructors

CoverageIndex

Fields

_coverageMetadata :: Doc ann

Decoder s [CoverageMetadata]

type Rep CoverageMetadata

Instance details

Defined in PlutusTx.Coverage

type Rep CoverageMetadata = D1 ('MetaData "CoverageMetadata" "PlutusTx.Coverage" "plutus-tx-1.60.0.0-inplace" 'True) (C1 ('MetaCons "CoverageMetadata" 'PrefixI 'True) (S1 ('MetaSel ('Just "_metadataSet") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 (Set Metadata))))

data Metadata

Constructors

ApplicationHeadSymbol String

IgnoredAnnotation
Location that is not interesting to cover. This is not generated by the compiler, but can be added later using addCoverageMetadata.

metadataSet :: Iso' CoverageMetadata (Set Metadata)

coverageAnnotations :: Getter CoverageIndex (Set CoverageAnnotation)

ignoredAnnotations :: Getter CoverageIndex (Set CoverageAnnotation)

coverageMetadata :: Iso' CoverageIndex (Map CoverageAnnotation CoverageMetadata)

coveredAnnotations :: Iso' CoverageData (Set CoverageAnnotation)

addCoverageMetadata :: CoverageAnnotation -> Metadata -> CoverageIndex -> CoverageIndex

Add metadata to a coverage annotation. Does nothing if the annotation is not in the index.

addLocationToCoverageIndex :: MonadWriter CoverageIndex m => CovLoc -> m CoverageAnnotation

Include a location coverage annotation in the index

addBoolCaseToCoverageIndex :: MonadWriter CoverageIndex m => CovLoc -> Bool -> CoverageMetadata -> m CoverageAnnotation

Include a boolean coverage annotation in the index

coverageDataFromLogMsg :: String -> CoverageData