PlutusTx.IsData.Class

Documentation

class ToData (a :: Type) where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData :: a -> BuiltinData

Convert a value to BuiltinData.

class FromData (a :: Type) where

Methods

fromBuiltinData :: BuiltinData -> Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData (a :: Type) where

Methods

unsafeFromBuiltinData :: BuiltinData -> a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

toData :: ToData a => a -> Data

Convert a value to Data.

fromData :: FromData a => Data -> Maybe a

Convert a value from Data, returning Nothing if this fails.