PlutusTx.Lift.Class

Documentation

class Typeable uni (a :: k) where

Class for types which have a corresponding Plutus IR type. Instances should always be derived, do not write your own instance!

Methods

typeRep :: Proxy a -> RTCompile uni fun (Type TyName uni ())

Get the Plutus IR type corresponding to this type.

class Lift uni a where

Class for types which can be lifted into Plutus IR. Instances should be derived, do not write your own instance!

Methods

lift :: a -> RTCompile uni fun (Term TyName Name uni fun ())

Get a Plutus IR term corresponding to the given value.

type RTCompile uni fun = DefT Name uni fun () Quote

makeTypeable :: Type -> Name -> Q [Dec]

makeLift :: Name -> Q [Dec]

withTyVars :: MonadReader (LocalVars uni) m => [(Name, TyVarDecl TyName ())] -> m a -> m a

data LiftError

Constructors

UnsupportedLiftKind Kind

UnsupportedLiftType Type

UserLiftError Text

LiftMissingDataCons Name

LiftMissingVar Name