PlutusTx.Plugin.Utils

Documentation

plc :: forall (loc :: Symbol) a. Proxy loc -> a -> CompiledCode a

Marks the given expression for compilation to PLC.

plinthc :: forall a. a -> CompiledCode a

anchor :: forall (loc :: Symbol) a. a -> a

This function is used in typeCheckResultAction to mark the given expression with its source location.

unsupported :: forall (err :: Symbol) (loc :: Symbol) a. a -> a

This function is used in typeCheckResultAction to mark the given expression as unsupported by Plinth.