PlutusTx.Prelude

Documentation

The PlutusTx Prelude is a replacement for the Haskell Prelude that works better with Plutus Tx. You should use it if you're writing code that will be compiled with the Plutus Tx compiler. {-# LANGUAGE NoImplicitPrelude #-} import PlutusTx.Prelude

module PlutusTx.Eq

module PlutusTx.Enum

module PlutusTx.Ord

module PlutusTx.Semigroup

module PlutusTx.Monoid

module PlutusTx.Numeric

module PlutusTx.Functor

module PlutusTx.Applicative

module PlutusTx.Lattice

module PlutusTx.Foldable

module PlutusTx.Traversable

Monad

(>>=) :: Monad m => m a -> (a -> m b) -> m b infixl 1

Sequentially compose two actions, passing any value produced by the first as an argument to the second.

'as >>= bs' can be understood as the do expression

do a <- as
   bs a

(=<<) :: Monad m => (a -> m b) -> m a -> m b infixr 1

Same as >>=, but with the arguments interchanged.

(>>) :: Monad m => m a -> m b -> m b infixl 1

Sequentially compose two actions, discarding any value produced by the first, like sequencing operators (such as the semicolon) in imperative languages.

'as >> bs' can be understood as the do expression

do as
   bs

return :: Monad m => a -> m a

Inject a value into the monadic type.

Standard functions, Tuples

module PlutusTx.Base

Tracing functions

module PlutusTx.Trace

String

data BuiltinString

appendString :: BuiltinString -> BuiltinString -> BuiltinString

Append two Strings.

emptyString :: BuiltinString

An empty String.

equalsString :: BuiltinString -> BuiltinString -> Bool

Check if two strings are equal

encodeUtf8 :: BuiltinString -> BuiltinByteString

Convert a String into a ByteString.

Error

error :: () -> a

Aborts evaluation with an error.

check :: Bool -> ()

Checks a Bool and aborts if it is false.

Booleans

module PlutusTx.Bool

Integer numbers

data Integer

Arbitrary precision integers. In contrast with fixed-size integral types such as Int, the Integer type represents the entire infinite range of integers.

For more information about this type's representation, see the comments in its implementation.

divide :: Integer -> Integer -> Integer

Integer division, rounding downwards

>>> divide (-41) 5
-9

modulo :: Integer -> Integer -> Integer

Integer remainder, always positive for a positive divisor

>>> modulo (-41) 5
4

quotient :: Integer -> Integer -> Integer

Integer division, rouding towards zero

>>> quotient (-41) 5
-8

remainder :: Integer -> Integer -> Integer

Integer remainder, same sign as dividend

>>> remainder (-41) 5
-1

even :: Integer -> Bool

Maybe

module PlutusTx.Maybe

Either

module PlutusTx.Either

Lists

map :: (a -> b) -> [a] -> [b]

Plutus Tx version of map.

>>> map (\i -> i + 1) [1, 2, 3]
[2,3,4]

(++) :: [a] -> [a] -> [a] infixr 5

Plutus Tx version of (++).

>>> [0, 1, 2] ++ [1, 2, 3, 4]
[0,1,2,1,2,3,4]

filter :: (a -> Bool) -> [a] -> [a]

Plutus Tx version of filter.

>>> filter (> 1) [1, 2, 3, 4]
[2,3,4]

listToMaybe :: [a] -> Maybe a

Plutus Tx version of listToMaybe.

uniqueElement :: [a] -> Maybe a

Return the element in the list, if there is precisely one.

findIndices :: (a -> Bool) -> [a] -> [Integer]

Plutus Tx version of findIndices.

findIndex :: (a -> Bool) -> [a] -> Maybe Integer

Plutus Tx version of findIndex.

(!!) :: [a] -> Integer -> a infixl 9

Plutus Tx version of (!!).

>>> [10, 11, 12] !! 2
12

reverse :: [a] -> [a]

Plutus Tx version of reverse.

zip :: [a] -> [b] -> [(a, b)]

Plutus Tx version of zip.

head :: [a] -> a

Plutus Tx version of head.

tail :: [a] -> [a]

Plutus Tx version of tail.

take :: Integer -> [a] -> [a]

Plutus Tx version of take.

nub :: Eq a => [a] -> [a]

Plutus Tx version of nub.

nubBy :: (a -> a -> Bool) -> [a] -> [a]

Plutus Tx version of nubBy.

zipWith :: (a -> b -> c) -> [a] -> [b] -> [c]

Plutus Tx version of zipWith.

dropWhile :: (a -> Bool) -> [a] -> [a]

Plutus Tx version of dropWhile.

partition :: (a -> Bool) -> [a] -> ([a], [a])

Plutus Tx version of partition.

sort :: Ord a => [a] -> [a]

Plutus Tx version of sort.

sortBy :: (a -> a -> Ordering) -> [a] -> [a]

Plutus Tx version of sortBy.

ByteStrings

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

appendByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Concatenates two ByteStrings.

consByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Adds a byte to the front of a ByteString.

takeByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Returns the n length prefix of a ByteString.

dropByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Returns the suffix of a ByteString after n elements.

sliceByteString :: Integer -> Integer -> BuiltinByteString -> BuiltinByteString

Returns the substring of a ByteString from index start of length n.

lengthOfByteString :: BuiltinByteString -> Integer

Returns the length of a ByteString.

indexByteString :: BuiltinByteString -> Integer -> Integer

Returns the byte of a ByteString at index.

emptyByteString :: BuiltinByteString

An empty ByteString.

decodeUtf8 :: BuiltinByteString -> BuiltinString

Converts a ByteString to a String.

Hashes and Signatures

sha2_256 :: BuiltinByteString -> BuiltinByteString

The SHA2-256 hash of a ByteString

sha3_256 :: BuiltinByteString -> BuiltinByteString

The SHA3-256 hash of a ByteString

verifyEd25519Signature

Arguments

:: BuiltinByteString
Public Key (32 bytes)

-> BuiltinByteString
Message (arbirtary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Ed25519 signature verification. Verify that the signature is a signature of the message by the public key. This will fail if key or the signature are not of the expected length.

verifyEcdsaSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (64 bytes)

-> BuiltinByteString
Message hash (32 bytes)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given an ECDSA SECP256k1 verification key, an ECDSA SECP256k1 signature, and an ECDSA SECP256k1 message hash (all as BuiltinByteStrings), verify the hash with that key and signature.

Important note

The verification key, the signature, and the message hash must all be of appropriate form and length. This function will error if any of these are not the case.

verifySchnorrSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (64 bytes)

-> BuiltinByteString
Message

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given a Schnorr SECP256k1 verification key, a Schnorr SECP256k1 signature, and a message (all as BuiltinByteStrings), verify the message with that key and signature.

Important note

The verification key and signature must all be of appropriate form and length. This function will error if this is not the case.

Rational numbers

data Rational

Represents an arbitrary-precision ratio.

unsafeRatio :: Integer -> Integer -> Rational

Makes a Rational from a numerator and a denominator.

Important note

If given a zero denominator, this function will error. If you don't mind a size increase, and care about safety, use ratio instead.

ratio :: Integer -> Integer -> Maybe Rational

Safely constructs a Rational from a numerator and a denominator. Returns Nothing if given a zero denominator.

fromInteger :: Integer -> Rational

Converts an Integer into the equivalent Rational.

round :: Rational -> Integer

round r returns the nearest Integer value to r. If r is equidistant between two values, the even value will be given.

Data

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

fromBuiltin :: FromBuiltin arep a => arep -> a

toBuiltin :: ToBuiltin a arep => a -> arep