PlutusTx

data CompiledCodeIn uni fun a

A compiled Plutus Tx program. The last type parameter indicates the type of the Haskell expression that was compiled, and hence the type of the compiled code.

Note: the compiled PLC program does *not* have normalized types, if you want to put it on the chain you must normalize the types first.

getPlc :: (SomeTypeIn uni)) => CompiledCodeIn uni fun (a -> b) -> CompiledCodeIn uni fun a -> CompiledCodeIn uni fun b

Apply a compiled function to a compiled argument. Will throw if the versions don't match, should only be used in non-production code.

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

HasToOpaque (BuiltinData, BuiltinData) (BuiltinPair BuiltinData BuiltinData)

Instance details

Defined in PlutusTx.Builtins.HasOpaque

Methods

toOpaque :: (BuiltinData, BuiltinData) -> BuiltinPair BuiltinData BuiltinData

type Rep BuiltinData

Instance details

Defined in PlutusTx.Builtins.Internal

type Rep BuiltinData = D1 ('MetaData "BuiltinData" "PlutusTx.Builtins.Internal" "plutus-tx-1.60.0.0-inplace" 'False) (C1 ('MetaCons "BuiltinData" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'SourceLazy 'DecidedLazy) (Rec0 Data)))

type Unroll BuiltinData

Instance details

Defined in PlutusTx.Blueprint.Definition.Unroll

type Unroll BuiltinData = '[BuiltinData]

type FromBuiltin BuiltinData

Instance details

Defined in PlutusTx.Builtins.HasBuiltin

type FromBuiltin BuiltinData = Data

data Data

A generic "data" type.

The main constructor Constr represents a datatype value in sum-of-products form: Constr i args represents a use of the ith constructor along with its arguments.

The other constructors are various primitives.

Constructors

Constr Integer [Data]

Map [(Data, Data)]

List [Data]

I Integer

B DefaultUni Data)

class ToData (a :: Type) where

A typeclass for types that can be converted to and from BuiltinData.

Methods

toBuiltinData :: a -> BuiltinData

Convert a value to BuiltinData.

class FromData (a :: Type) where

Methods

fromBuiltinData :: BuiltinData -> Maybe a

Convert a value from BuiltinData, returning Nothing if this fails.

class UnsafeFromData (a :: Type) where

Methods

unsafeFromBuiltinData :: BuiltinData -> a

Convert a value from BuiltinData, calling error if this fails. This is typically much faster than fromBuiltinData.

When implementing this function, make sure to call unsafeFromBuiltinData rather than fromBuiltinData when converting substructures!

This is a simple type without any validation, use with caution.

toData :: ToData a => a -> Data

Convert a value to Data.

Note: off-chain only.

fromData :: FromData a => Data -> Maybe a

Convert a value from Data, returning Nothing if this fails.

Note: off-chain only.

builtinDataToData :: BuiltinData -> Data

NOT a builtin. Converts a BuiltinData into a Data. Only works off-chain.

dataToBuiltinData :: Data -> BuiltinData

NOT a builtin. Converts a Data into a BuiltinData. Only works off-chain.

unstableMakeIsData :: Name -> Q [Dec]

Generate a FromData and a ToData instance for a type. This may not be stable in the face of constructor additions, renamings, etc. Use makeIsDataIndexed if you need stability.

makeIsDataIndexed :: Name -> [(Name, Int)] -> Q [Dec]

Generate a ToData, 'FromData and a UnsafeFromData instances for a type, using an explicit mapping of constructor names to indices. Use this for types where you need to keep the representation stable.

makeIsDataSchemaIndexed :: Name -> [(Name, Natural)] -> Q [InstanceDec]

Generate a ToData, FromData, UnsafeFromData, HasBlueprintSchema instances for a type, using an explicit mapping of constructor names to indices. Use this for types where you need to keep the representation stable.

Note: Requires ViewPatterns, FlexibleInstances, UndecidableInstances, MultiParamTypeClasses, TypeApplications language extensions and an `import PlutusTx.Blueprint.Definition (definitionRef)`.

class Lift uni a

Class for types which can be lifted into Plutus IR. Instances should be derived, do not write your own instance!

Minimal complete definition

lift

class Typeable uni (a :: k)

Class for types which have a corresponding Plutus IR type. Instances should always be derived, do not write your own instance!

Minimal complete definition

typeRep