PlutusTx.Builtins

Description

Primitive names and functions for working with Plutus Core builtins.

Bytestring builtins

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

appendByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Concatenates two ByteStrings.

consByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Adds a byte to the front of a ByteString.

sliceByteString :: Integer -> Integer -> BuiltinByteString -> BuiltinByteString

Returns the substring of a ByteString from index start of length n.

lengthOfByteString :: BuiltinByteString -> Integer

Returns the length of a ByteString.

indexByteString :: BuiltinByteString -> Integer -> Integer

Returns the byte of a ByteString at index.

emptyByteString :: BuiltinByteString

An empty ByteString.

equalsByteString :: BuiltinByteString -> BuiltinByteString -> Bool

Check if two ByteStrings are equal.

lessThanByteString :: BuiltinByteString -> BuiltinByteString -> Bool

Check if one ByteString is less than another.

lessThanEqualsByteString :: BuiltinByteString -> BuiltinByteString -> Bool

Check if one ByteString is less than or equal to another.

greaterThanByteString :: BuiltinByteString -> BuiltinByteString -> Bool

Check if one ByteString is greater than another.

greaterThanEqualsByteString :: BuiltinByteString -> BuiltinByteString -> Bool

Check if one ByteString is greater than another.

sha2_256 :: BuiltinByteString -> BuiltinByteString

The SHA2-256 hash of a ByteString

sha3_256 :: BuiltinByteString -> BuiltinByteString

The SHA3-256 hash of a ByteString

blake2b_224 :: BuiltinByteString -> BuiltinByteString

The BLAKE2B-224 hash of a ByteString

blake2b_256 :: BuiltinByteString -> BuiltinByteString

The BLAKE2B-256 hash of a ByteString

keccak_256 :: BuiltinByteString -> BuiltinByteString

The KECCAK-256 hash of a ByteString

verifyEd25519Signature

Arguments

:: BuiltinByteString
Public Key (32 bytes)

-> BuiltinByteString
Message (arbirtary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Ed25519 signature verification. Verify that the signature is a signature of the message by the public key. This will fail if key or the signature are not of the expected length.

verifyEcdsaSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (33 bytes)

-> BuiltinByteString
Message hash (32 bytes)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given an ECDSA SECP256k1 verification key, an ECDSA SECP256k1 signature, and an ECDSA SECP256k1 message hash (all as BuiltinByteStrings), verify the hash with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length:

The first byte of the public key must correspond to the sign of the y coordinate: this is 0x02 if y is even, and 0x03 otherwise.

The remaining bytes of the public key must correspond to the x coordinate, as a big-endian integer.

The first 32 bytes of the signature must correspond to the big-endian integer representation of _r_.

The last 32 bytes of the signature must correspond to the big-endian integer representation of _s_.

While this primitive accepts a hash, any caller should only pass it hashes that they computed themselves: specifically, they should receive the message from a sender and hash it, rather than receiving the hash from said sender. Failure to do so can be dangerous. Other than length, we make no requirements of what hash gets used.

See also

secp256k1_ec_pubkey_serialize; this implements the format for the verification key that we accept, given a length argument of 33 and the SECP256K1_EC_COMPRESSED flag.

secp256k1_ecdsa_serialize_compact; this implements the format for the signature that we accept.

verifySchnorrSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (32 bytes)

-> BuiltinByteString
Message (arbitrary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given a Schnorr SECP256k1 verification key, a Schnorr SECP256k1 signature, and a message (all as BuiltinByteStrings), verify the message with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length. Throughout, we refer to co-ordinates of the point R.

The bytes of the public key must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The first 32 bytes of the signature must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The last 32 bytes of the signature must correspond to the bytes of s, as a big-endian integer, as specified in BIP-340.

See also

BIP-340

secp256k1_xonly_pubkey_serialize; this implements the format for the verification key that we accept.

secp256k1_schnorrsig_sign; this implements the signing logic for signatures this builtin can verify.

decodeUtf8 :: BuiltinByteString -> BuiltinString

Converts a ByteString to a String.

Integer builtins

data Integer

Arbitrary precision integers. In contrast with fixed-size integral types such as Int, the Integer type represents the entire infinite range of integers.

Integers are stored in a kind of sign-magnitude form, hence do not expect two's complement form when using bit operations.

If the value is small (fit into an Int), IS constructor is used. Otherwise Integer and IN constructors are used to store a BigNat representing respectively the positive or the negative value magnitude.

Invariant: Integer and IN are used iff value doesn't fit in IS

addInteger :: Integer -> Integer -> Integer

Add two Integers.

subtractInteger :: Integer -> Integer -> Integer

Subtract two Integers.

multiplyInteger :: Integer -> Integer -> Integer

Multiply two Integers.

divideInteger :: Integer -> Integer -> Integer

Divide two integers.

modInteger :: Integer -> Integer -> Integer

Integer modulo operation.

quotientInteger :: Integer -> Integer -> Integer

Quotient of two integers.

remainderInteger :: Integer -> Integer -> Integer

Take the remainder of dividing two Integers.

greaterThanInteger :: Integer -> Integer -> Bool

Check whether one Integer is greater than another.

greaterThanEqualsInteger :: Integer -> Integer -> Bool

Check whether one Integer is greater than or equal to another.

lessThanInteger :: Integer -> Integer -> Bool

Check whether one Integer is less than another.

lessThanEqualsInteger :: Integer -> Integer -> Bool

Check whether one Integer is less than or equal to another.

equalsInteger :: Integer -> Integer -> Bool

Check if two Integers are equal.

Error

error :: () -> a

Aborts evaluation with an error.

Data

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

chooseData :: forall a. BuiltinData -> a -> a -> a -> a -> a -> a

Given five values for the five different constructors of BuiltinData, selects one depending on which corresponds to the actual constructor of the given value.

matchData :: BuiltinData -> (Integer -> [BuiltinData] -> r) -> ([(BuiltinData, BuiltinData)] -> r) -> ([BuiltinData] -> r) -> (Integer -> r) -> (BuiltinByteString -> r) -> r

Given a BuiltinData value and matching functions for the five constructors, applies the appropriate matcher to the arguments of the constructor and returns the result.

matchData' :: BuiltinData -> (Integer -> BuiltinList BuiltinData -> r) -> (BuiltinList (BuiltinPair BuiltinData BuiltinData) -> r) -> (BuiltinList BuiltinData -> r) -> (Integer -> r) -> (BuiltinByteString -> r) -> r

Given a BuiltinData value and matching functions for the five constructors, applies the appropriate matcher to the arguments of the constructor and returns the result.

equalsData :: BuiltinData -> BuiltinData -> Bool

Check if two BuiltinDatas are equal.

serialiseData :: BuiltinData -> BuiltinByteString

Convert a String into a ByteString.

mkConstr :: Integer -> [BuiltinData] -> BuiltinData

Constructs a BuiltinData value with the Constr constructor.

mkMap :: [(BuiltinData, BuiltinData)] -> BuiltinData

Constructs a BuiltinData value with the Map constructor.

mkList :: [BuiltinData] -> BuiltinData

Constructs a BuiltinData value with the List constructor.

mkI :: Integer -> BuiltinData

Constructs a BuiltinData value with the I constructor.

mkB :: BuiltinByteString -> BuiltinData

Constructs a BuiltinData value with the B constructor.

unsafeDataAsConstr :: BuiltinData -> (Integer, [BuiltinData])

Deconstructs a BuiltinData as a Constr, or fails if it is not one.

unsafeDataAsMap :: BuiltinData -> [(BuiltinData, BuiltinData)]

Deconstructs a BuiltinData as a Map, or fails if it is not one.

unsafeDataAsList :: BuiltinData -> [BuiltinData]

Deconstructs a BuiltinData as a List, or fails if it is not one.

unsafeDataAsI :: BuiltinData -> Integer

Deconstructs a BuiltinData as an I, or fails if it is not one.

unsafeDataAsB :: BuiltinData -> BuiltinByteString

Deconstructs a BuiltinData as a B, or fails if it is not one.

builtinDataToData :: BuiltinData -> Data

Convert a BuiltinData into a Data. Only works off-chain.

dataToBuiltinData :: Data -> BuiltinData

Convert a Data into a BuiltinData. Only works off-chain.

Strings

data BuiltinString

appendString :: BuiltinString -> BuiltinString -> BuiltinString

Append two Strings.

emptyString :: BuiltinString

An empty String.

equalsString :: BuiltinString -> BuiltinString -> Bool

Check if two strings are equal

encodeUtf8 :: BuiltinString -> BuiltinByteString

Convert a String into a ByteString.

Pairs

pairToPair :: BuiltinPair a b -> (a, b)

Turn a builtin pair into a normal pair, useful in patterns.

Lists

matchList :: forall a r. BuiltinList a -> r -> (a -> BuiltinList a -> r) -> r

headMaybe :: BuiltinList a -> Maybe a

head :: BuiltinList a -> a

tail :: BuiltinList a -> BuiltinList a

uncons :: BuiltinList a -> Maybe (a, BuiltinList a)

Uncons a builtin list, failing if the list is empty, useful in patterns.

unsafeUncons :: BuiltinList a -> (a, BuiltinList a)

Uncons a builtin list, failing if the list is empty, useful in patterns.

Tracing

trace :: BuiltinString -> a -> a

Emit the given string as a trace message before evaluating the argument.

BLS12_381

data BuiltinBLS12_381_G1_Element

bls12_381_G1_equals :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element -> Bool

bls12_381_G1_add :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_scalarMul :: Integer -> BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_neg :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_compress :: BuiltinBLS12_381_G1_Element -> BuiltinByteString

bls12_381_G1_uncompress :: BuiltinByteString -> BuiltinBLS12_381_G1_Element

bls12_381_G1_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G1_Element

bls12_381_G1_compressed_zero :: BuiltinByteString

bls12_381_G1_compressed_generator :: BuiltinByteString

data BuiltinBLS12_381_G2_Element

bls12_381_G2_equals :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element -> Bool

bls12_381_G2_add :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_scalarMul :: Integer -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_neg :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_compress :: BuiltinBLS12_381_G2_Element -> BuiltinByteString

bls12_381_G2_uncompress :: BuiltinByteString -> BuiltinBLS12_381_G2_Element

bls12_381_G2_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G2_Element

bls12_381_G2_compressed_zero :: BuiltinByteString

bls12_381_G2_compressed_generator :: BuiltinByteString

data BuiltinBLS12_381_MlResult

bls12_381_millerLoop :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_MlResult

bls12_381_mulMlResult :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult

bls12_381_finalVerify :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> Bool

Conversions

fromBuiltin :: FromBuiltin arep a => arep -> a

toBuiltin :: ToBuiltin a arep => a -> arep

integerToByteString :: ByteOrder -> Integer -> Integer -> BuiltinByteString

Convert a BuiltinInteger into a BuiltinByteString, as described in CIP-0087. The first argument indicates the endianness of the conversion and the third argument is the integer to be converted, which must be non-negative. The second argument must also be non-negative and it indicates the required width of the output. If the width is zero then the output is the smallest bytestring which can contain the converted input (and in this case, the integer 0 encodes to the empty bytestring). If the width is nonzero then the output bytestring will be padded to the required width with 0x00 bytes (on the left for big-endian conversions and on the right for little-endian conversions); if the input integer is too big to fit into a bytestring of the specified width then the conversion will fail. Conversion will also fail if the specified width is greater than 8192 or the input integer is too big to fit into a bytestring of length 8192.

byteStringToInteger :: ByteOrder -> BuiltinByteString -> Integer

Convert a BuiltinByteString to a BuiltinInteger, as described in CIP-0087. The first argument indicates the endianness of the conversion and the second is the bytestring to be converted. There is no limitation on the size of the bytestring. The empty bytestring is converted to the integer 0.