PlutusTx.Builtins.Class

Documentation

class FromBuiltin arep a | arep -> a where

A class witnessing the ability to convert from the builtin representation to the Haskell representation.

Methods

fromBuiltin :: arep -> a

class ToBuiltin a arep | a -> arep where

A class witnessing the ability to convert from the Haskell representation to the builtin representation.

Methods

toBuiltin :: a -> arep

stringToBuiltinString :: String -> BuiltinString

obfuscatedId :: a -> a

stringToBuiltinByteString :: String -> BuiltinByteString

Orphan instances

IsString BuiltinByteString

Instance details

Methods

fromString :: String -> BuiltinByteString

IsString BuiltinString

Instance details

Methods

fromString :: String -> BuiltinString