PlutusTx.Builtins.Internal

Description

This module contains the special Haskell names that are used to map to builtin types or functions in Plutus Core.

Most users should not use this module directly, but rather use Builtins.

Documentation

error :: BuiltinUnit -> a

data BuiltinBool

Constructors

BuiltinBool Bool

true :: BuiltinBool

false :: BuiltinBool

ifThenElse :: BuiltinBool -> a -> a -> a

data BuiltinUnit

Constructors

BuiltinUnit ()

unitval :: BuiltinUnit

chooseUnit :: BuiltinUnit -> a -> a

type BuiltinInteger = Integer

addInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

subtractInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

multiplyInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

divideInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

modInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

quotientInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

remainderInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinInteger

lessThanInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinBool

lessThanEqualsInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinBool

equalsInteger :: BuiltinInteger -> BuiltinInteger -> BuiltinBool

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

Constructors

BuiltinByteString ByteString

appendByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString

consByteString :: BuiltinInteger -> BuiltinByteString -> BuiltinByteString

sliceByteString :: BuiltinInteger -> BuiltinInteger -> BuiltinByteString -> BuiltinByteString

lengthOfByteString :: BuiltinByteString -> BuiltinInteger

indexByteString :: BuiltinByteString -> BuiltinInteger -> BuiltinInteger

emptyByteString :: BuiltinByteString

sha2_256 :: BuiltinByteString -> BuiltinByteString

sha3_256 :: BuiltinByteString -> BuiltinByteString

blake2b_256 :: BuiltinByteString -> BuiltinByteString

verifyEd25519Signature :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString -> BuiltinBool

verifyEcdsaSecp256k1Signature :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString -> BuiltinBool

verifySchnorrSecp256k1Signature :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString -> BuiltinBool

traceAll :: forall (a :: Type) (f :: Type -> Type). Foldable f => f Text -> a -> a

equalsByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinBool

lessThanByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinBool

lessThanEqualsByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinBool

decodeUtf8 :: BuiltinByteString -> BuiltinString

data BuiltinString

Constructors

BuiltinString Text

appendString :: BuiltinString -> BuiltinString -> BuiltinString

emptyString :: BuiltinString

equalsString :: BuiltinString -> BuiltinString -> BuiltinBool

trace :: BuiltinString -> a -> a

encodeUtf8 :: BuiltinString -> BuiltinByteString

data BuiltinPair a b

Constructors

BuiltinPair (a, b)

fst :: BuiltinPair a b -> a

snd :: BuiltinPair a b -> b

mkPairData :: BuiltinData -> BuiltinData -> BuiltinPair BuiltinData BuiltinData

data BuiltinList a

Constructors

BuiltinList [a]

null :: BuiltinList a -> BuiltinBool

head :: BuiltinList a -> a

tail :: BuiltinList a -> BuiltinList a

chooseList :: BuiltinList a -> b -> b -> b

mkNilData :: BuiltinUnit -> BuiltinList BuiltinData

mkNilPairData :: BuiltinUnit -> BuiltinList (BuiltinPair BuiltinData BuiltinData)

mkCons :: a -> BuiltinList a -> BuiltinList a

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

Constructors

BuiltinData Data

builtinDataToData :: BuiltinData -> Data

Convert a BuiltinData into a Data. Only works off-chain.

dataToBuiltinData :: Data -> BuiltinData

Convert a Data into a BuiltinData. Only works off-chain.

chooseData :: forall a. BuiltinData -> a -> a -> a -> a -> a -> a

mkConstr :: BuiltinInteger -> BuiltinList BuiltinData -> BuiltinData

mkMap :: BuiltinList (BuiltinPair BuiltinData BuiltinData) -> BuiltinData

mkList :: BuiltinList BuiltinData -> BuiltinData

mkI :: BuiltinInteger -> BuiltinData

mkB :: BuiltinByteString -> BuiltinData

unsafeDataAsConstr :: BuiltinData -> BuiltinPair BuiltinInteger (BuiltinList BuiltinData)

unsafeDataAsMap :: BuiltinData -> BuiltinList (BuiltinPair BuiltinData BuiltinData)

unsafeDataAsList :: BuiltinData -> BuiltinList BuiltinData

unsafeDataAsI :: BuiltinData -> BuiltinInteger

unsafeDataAsB :: BuiltinData -> BuiltinByteString

equalsData :: BuiltinData -> BuiltinData -> BuiltinBool

serialiseData :: BuiltinData -> BuiltinByteString