PlutusTx.Code

Documentation

data CompiledCodeIn uni fun ann a

A compiled Plutus Tx program. The last type parameter indicates the type of the Haskell expression that was compiled, and hence the type of the compiled code.

Note: the compiled PLC program does *not* have normalized types, if you want to put it on the chain you must normalize the types first.

Constructors

SerializedCode ByteString (Maybe ByteString) CoverageIndex
Serialized UPLC code and possibly serialized PIR code with metadata used for program coverage.

DeserializedCode (Program NamedDeBruijn uni fun ann) (Maybe (Program TyName Name uni fun ann)) CoverageIndex
Deserialized UPLC program, and possibly deserialized PIR program with metadata used for program coverage.

type CompiledCode = CompiledCodeIn DefaultUni DefaultFun ()

CompiledCodeIn instantiated with default built-in types and functions, and empty annotation.

data SrcSpan

The span between two source locations.

This corresponds roughly to the SrcSpan used by GHC, but we define our own version so we don't have to depend on ghc to use it.

The line and column numbers are 1-based, and the unit is Unicode code point (or Char).

Constructors

SrcSpan

Fields

srcSpanFile :: FilePath

srcSpanSLine :: Int

srcSpanSCol :: Int

srcSpanELine :: Int

srcSpanECol :: Int

type SrcSpans = Set SrcSpan

type CompiledCodeDebug = CompiledCodeIn DefaultUni DefaultFun SrcSpans

CompiledCodeIn instantiated with default built-in types and functions, and a set of SrcSpans as annotation.

applyCode :: (Closed uni, uni `Everywhere` Flat, Flat fun) => CompiledCodeIn uni fun () (a -> b) -> CompiledCodeIn uni fun () a -> CompiledCodeIn uni fun () b

Apply a compiled function to a compiled argument.

sizePlc :: (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann) => CompiledCodeIn uni fun ann a -> Integer

The size of a CompiledCodeIn, in AST nodes.

newtype ImpossibleDeserialisationFailure

Constructors

ImpossibleDeserialisationFailure DecodeException

getPlc :: (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann) => CompiledCodeIn uni fun ann a -> Program NamedDeBruijn uni fun ann

Get the actual Plutus Core program out of a CompiledCodeIn.

getPir :: (Closed uni, uni `Everywhere` Flat, Flat fun, Flat ann) => CompiledCodeIn uni fun ann a -> Maybe (Program TyName Name uni fun ann)

Get the Plutus IR program, if there is one, out of a CompiledCodeIn.

getCovIdx :: CompiledCodeIn uni fun ann a -> CoverageIndex