PlutusTx.Lift

Get a Plutus Core term corresponding to the given value. Allows configuring PIR and UPLC optimization options.

safeLift :: forall a uni fun m. (Lift uni a, NamedDeBruijn uni fun ())

Get a Plutus Core term corresponding to the given value, applying default PIR/UPLC optimizations.

safeLiftUnopt :: forall a uni fun m. (Lift uni a, NamedDeBruijn uni fun ())

Like safeLift but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

safeLiftProgram :: (Lift uni a, NamedDeBruijn uni fun ())

Get a Plutus Core program corresponding to the given value, applying default PIR/UPLC optimizations.

safeLiftProgramUnopt :: (Lift uni a, NamedDeBruijn uni fun ())

Like safeLiftProgram but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

safeLiftCode :: (Lift uni a, Version -> a -> m (CompiledCodeIn uni fun a)

Like safeLiftCode but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

lift :: (Lift uni a, NamedDeBruijn uni fun ())

Get a Plutus Core term corresponding to the given value, throwing any errors that occur as exceptions and ignoring fresh names. The default PIR/UPLC optimizations are applied.

liftUnopt :: (Lift uni a, NamedDeBruijn uni fun ())

Like lift but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftProgram :: (Lift uni a, NamedDeBruijn uni fun ())

Like liftProgram but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftProgramDef :: Lift DefaultFun ())

Like liftProgramDef but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftCode :: (Lift uni a, Version -> a -> CompiledCodeIn uni fun a

Like liftCode but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

liftCodeDef :: (Lift uni a, RewriteRules uni fun), Hashable fun) => a -> CompiledCodeIn uni fun a

Like liftCodeDef but does not apply PIR/UPLC optimizations. Use this option where lifting speed is more important than optimal code.

typeCheckAgainst :: forall a uni fun m. (Typeable uni a, Doc ann