PlutusTx.Prelude

Documentation

The PlutusTx Prelude is a replacement for the Haskell Prelude that works better with Plutus Tx. You should use it if you're writing code that will be compiled with the Plutus Tx compiler. {-# LANGUAGE NoImplicitPrelude #-} import PlutusTx.Prelude

module PlutusTx.Eq

module PlutusTx.Enum

module PlutusTx.Ord

module PlutusTx.Semigroup

module PlutusTx.Monoid

module PlutusTx.Numeric

module PlutusTx.Functor

module PlutusTx.Applicative

module PlutusTx.Lattice

module PlutusTx.Foldable

module PlutusTx.Traversable

Monad

(>>=) :: Monad m => m a -> (a -> m b) -> m b infixl 1

Sequentially compose two actions, passing any value produced by the first as an argument to the second.

'as >>= bs' can be understood as the do expression

do a <- as
   bs a

(=<<) :: Monad m => (a -> m b) -> m a -> m b infixr 1

Same as >>=, but with the arguments interchanged.

(>>) :: Monad m => m a -> m b -> m b infixl 1

Sequentially compose two actions, discarding any value produced by the first, like sequencing operators (such as the semicolon) in imperative languages.

'as >> bs' can be understood as the do expression

do as
   bs

return :: Monad m => a -> m a

Inject a value into the monadic type.

Standard functions, Tuples

module PlutusTx.Base

Tracing functions

module PlutusTx.Trace

String

data BuiltinString

appendString :: BuiltinString -> BuiltinString -> BuiltinString

Append two Strings.

emptyString :: BuiltinString

An empty String.

equalsString :: BuiltinString -> BuiltinString -> Bool

Check if two strings are equal

encodeUtf8 :: BuiltinString -> BuiltinByteString

Convert a String into a ByteString.

Error

error :: () -> a

Aborts evaluation with an error.

check :: Bool -> ()

Checks a Bool and aborts if it is false.

Booleans

module PlutusTx.Bool

Integer numbers

data Integer

Arbitrary precision integers. In contrast with fixed-size integral types such as Int, the Integer type represents the entire infinite range of integers.

Integers are stored in a kind of sign-magnitude form, hence do not expect two's complement form when using bit operations.

If the value is small (fit into an Int), IS constructor is used. Otherwise Integer and IN constructors are used to store a BigNat representing respectively the positive or the negative value magnitude.

Invariant: Integer and IN are used iff value doesn't fit in IS

divide :: Integer -> Integer -> Integer

Integer division, rounding downwards

>>> divide (-41) 5
-9

modulo :: Integer -> Integer -> Integer

Integer remainder, always positive for a positive divisor

>>> modulo (-41) 5
4

quotient :: Integer -> Integer -> Integer

Integer division, rouding towards zero

>>> quotient (-41) 5
-8

remainder :: Integer -> Integer -> Integer

Integer remainder, same sign as dividend

>>> remainder (-41) 5
-1

even :: Integer -> Bool

odd :: Integer -> Bool

Maybe

module PlutusTx.Maybe

Either

module PlutusTx.Either

Lists

zipWith :: forall a b c. (a -> b -> c) -> [a] -> [b] -> [c]

Plutus Tx version of zipWith.

sortBy :: (a -> a -> Ordering) -> [a] -> [a]

Plutus Tx version of sortBy.

head :: [a] -> a

Plutus Tx version of head.

filter :: (a -> Bool) -> [a] -> [a]

Plutus Tx version of filter.

>>> filter (> 1) [1, 2, 3, 4]
[2,3,4]

(++) :: [a] -> [a] -> [a] infixr 5

Plutus Tx version of (++).

>>> [0, 1, 2] ++ [1, 2, 3, 4]
[0,1,2,1,2,3,4]

zip :: [a] -> [b] -> [(a, b)]

Plutus Tx version of zip.

map :: forall a b. (a -> b) -> [a] -> [b]

Plutus Tx version of map.

>>> map (\i -> i + 1) [1, 2, 3]
[2,3,4]

listToMaybe :: [a] -> Maybe a

Plutus Tx version of listToMaybe.

uncons :: [a] -> Maybe (a, [a])

Plutus Tx version of uncons.

tail :: [a] -> [a]

Plutus Tx version of tail.

last :: [a] -> a

Plutus Tx version of last.

null :: [a] -> Bool

Test whether a list is empty.

replicate :: forall a. Integer -> a -> [a]

Plutus Tx version of replicate.

dropWhile :: forall a. (a -> Bool) -> [a] -> [a]

Plutus Tx version of dropWhile.

take :: Integer -> [a] -> [a]

Plutus Tx version of take.

drop :: Integer -> [a] -> [a]

Plutus Tx version of drop.

splitAt :: Integer -> [a] -> ([a], [a])

Plutus Tx version of splitAt.

reverse :: [a] -> [a]

Plutus Tx version of reverse.

and :: [Bool] -> Bool

Returns the conjunction of a list of Bools.

or :: [Bool] -> Bool

Returns the disjunction of a list of Bools.

any :: forall a. (a -> Bool) -> [a] -> Bool

Determines whether any element of the structure satisfies the predicate.

all :: forall a. (a -> Bool) -> [a] -> Bool

Determines whether all elements of the list satisfy the predicate.

elem :: Eq a => a -> [a] -> Bool

Does the element occur in the list?

notElem :: Eq a => a -> [a] -> Bool

The negation of elem.

(!!) :: forall a. [a] -> Integer -> a infixl 9

Plutus Tx version of (!!).

>>> [10, 11, 12] !! 2
12

unzip :: [(a, b)] -> ([a], [b])

Plutus Tx version of unzip.

find :: forall a. (a -> Bool) -> [a] -> Maybe a

Returns the leftmost element matching the predicate, or Nothing if there's no such element.

findIndex :: (a -> Bool) -> [a] -> Maybe Integer

Plutus Tx version of findIndex.

findIndices :: (a -> Bool) -> [a] -> [Integer]

Plutus Tx version of findIndices.

nub :: Eq a => [a] -> [a]

Plutus Tx version of nub.

nubBy :: (a -> a -> Bool) -> [a] -> [a]

Plutus Tx version of nubBy.

partition :: (a -> Bool) -> [a] -> ([a], [a])

Plutus Tx version of partition.

sort :: Ord a => [a] -> [a]

Plutus Tx version of sort.

uniqueElement :: [a] -> Maybe a

Return the element in the list, if there is precisely one.

revAppend :: forall a. [a] -> [a] -> [a]

Cons each element of the first list to the second one in reverse order (i.e. the last element of the first list is the head of the result).

revAppend xs ys === reverse xs ++ ys

>>> revAppend "abc" "de"
"cbade"

ByteStrings

data BuiltinByteString

An opaque type representing Plutus Core ByteStrings.

appendByteString :: BuiltinByteString -> BuiltinByteString -> BuiltinByteString

Concatenates two ByteStrings.

consByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Adds a byte to the front of a ByteString.

takeByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Returns the n length prefix of a ByteString.

dropByteString :: Integer -> BuiltinByteString -> BuiltinByteString

Returns the suffix of a ByteString after n elements.

sliceByteString :: Integer -> Integer -> BuiltinByteString -> BuiltinByteString

Returns the substring of a ByteString from index start of length n.

lengthOfByteString :: BuiltinByteString -> Integer

Returns the length of a ByteString.

indexByteString :: BuiltinByteString -> Integer -> Integer

Returns the byte of a ByteString at index.

emptyByteString :: BuiltinByteString

An empty ByteString.

decodeUtf8 :: BuiltinByteString -> BuiltinString

Converts a ByteString to a String.

Hashes and Signatures

sha2_256 :: BuiltinByteString -> BuiltinByteString

The SHA2-256 hash of a ByteString

sha3_256 :: BuiltinByteString -> BuiltinByteString

The SHA3-256 hash of a ByteString

blake2b_224 :: BuiltinByteString -> BuiltinByteString

The BLAKE2B-224 hash of a ByteString

blake2b_256 :: BuiltinByteString -> BuiltinByteString

The BLAKE2B-256 hash of a ByteString

keccak_256 :: BuiltinByteString -> BuiltinByteString

The KECCAK-256 hash of a ByteString

verifyEd25519Signature

Arguments

:: BuiltinByteString
Public Key (32 bytes)

-> BuiltinByteString
Message (arbirtary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Ed25519 signature verification. Verify that the signature is a signature of the message by the public key. This will fail if key or the signature are not of the expected length.

verifyEcdsaSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (33 bytes)

-> BuiltinByteString
Message hash (32 bytes)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given an ECDSA SECP256k1 verification key, an ECDSA SECP256k1 signature, and an ECDSA SECP256k1 message hash (all as BuiltinByteStrings), verify the hash with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length:

The first byte of the public key must correspond to the sign of the y coordinate: this is 0x02 if y is even, and 0x03 otherwise.

The remaining bytes of the public key must correspond to the x coordinate, as a big-endian integer.

The first 32 bytes of the signature must correspond to the big-endian integer representation of _r_.

The last 32 bytes of the signature must correspond to the big-endian integer representation of _s_.

While this primitive accepts a hash, any caller should only pass it hashes that they computed themselves: specifically, they should receive the message from a sender and hash it, rather than receiving the hash from said sender. Failure to do so can be dangerous. Other than length, we make no requirements of what hash gets used.

See also

secp256k1_ec_pubkey_serialize; this implements the format for the verification key that we accept, given a length argument of 33 and the SECP256K1_EC_COMPRESSED flag.

secp256k1_ecdsa_serialize_compact; this implements the format for the signature that we accept.

verifySchnorrSecp256k1Signature

Arguments

:: BuiltinByteString
Verification key (32 bytes)

-> BuiltinByteString
Message (arbitrary length)

-> BuiltinByteString
Signature (64 bytes)

-> Bool

Given a Schnorr SECP256k1 verification key, a Schnorr SECP256k1 signature, and a message (all as BuiltinByteStrings), verify the message with that key and signature.

Note

There are additional well-formation requirements for the arguments beyond their length. Throughout, we refer to co-ordinates of the point R.

The bytes of the public key must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The first 32 bytes of the signature must correspond to the x coordinate, as a big-endian integer, as specified in BIP-340.

The last 32 bytes of the signature must correspond to the bytes of s, as a big-endian integer, as specified in BIP-340.

See also

BIP-340

secp256k1_xonly_pubkey_serialize; this implements the format for the verification key that we accept.

secp256k1_schnorrsig_sign; this implements the signing logic for signatures this builtin can verify.

Rational numbers

data Rational

Represents an arbitrary-precision ratio.

The following two invariants are maintained:

The denominator is greater than zero.

The numerator and denominator are coprime.

unsafeRatio :: Integer -> Integer -> Rational

Makes a Rational from a numerator and a denominator.

Important note

If given a zero denominator, this function will error. If you don't mind a size increase, and care about safety, use ratio instead.

ratio :: Integer -> Integer -> Maybe Rational

Safely constructs a Rational from a numerator and a denominator. Returns Nothing if given a zero denominator.

fromInteger :: Integer -> Rational

Converts an Integer into the equivalent Rational.

round :: Rational -> Integer

round r returns the nearest Integer value to r. If r is equidistant between two values, the even value will be given.

Data

data BuiltinData

A type corresponding to the Plutus Core builtin equivalent of Data.

The point of this type is to be an opaque equivalent of Data, so as to ensure that it is only used in ways that the compiler can handle.

As such, you should use this type in your on-chain code, and in any data structures that you want to be representable on-chain.

For off-chain usage, there are conversion functions builtinDataToData and dataToBuiltinData, but note that these will not work on-chain.

BLS12_381

data BuiltinBLS12_381_G1_Element

bls12_381_G1_equals :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element -> Bool

bls12_381_G1_add :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_neg :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_scalarMul :: Integer -> BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G1_Element

bls12_381_G1_compress :: BuiltinBLS12_381_G1_Element -> BuiltinByteString

bls12_381_G1_uncompress :: BuiltinByteString -> BuiltinBLS12_381_G1_Element

bls12_381_G1_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G1_Element

bls12_381_G1_compressed_zero :: BuiltinByteString

bls12_381_G1_compressed_generator :: BuiltinByteString

data BuiltinBLS12_381_G2_Element

bls12_381_G2_equals :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element -> Bool

bls12_381_G2_add :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_neg :: BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_scalarMul :: Integer -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_G2_Element

bls12_381_G2_compress :: BuiltinBLS12_381_G2_Element -> BuiltinByteString

bls12_381_G2_uncompress :: BuiltinByteString -> BuiltinBLS12_381_G2_Element

bls12_381_G2_hashToGroup :: BuiltinByteString -> BuiltinByteString -> BuiltinBLS12_381_G2_Element

bls12_381_G2_compressed_zero :: BuiltinByteString

bls12_381_G2_compressed_generator :: BuiltinByteString

data BuiltinBLS12_381_MlResult

bls12_381_millerLoop :: BuiltinBLS12_381_G1_Element -> BuiltinBLS12_381_G2_Element -> BuiltinBLS12_381_MlResult

bls12_381_mulMlResult :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult

bls12_381_finalVerify :: BuiltinBLS12_381_MlResult -> BuiltinBLS12_381_MlResult -> Bool

byteStringToInteger :: ByteOrder -> BuiltinByteString -> Integer

Convert a BuiltinByteString to a BuiltinInteger, as described in CIP-0087. The first argument indicates the endianness of the conversion and the second is the bytestring to be converted. There is no limitation on the size of the bytestring. The empty bytestring is converted to the integer 0.

integerToByteString :: ByteOrder -> Integer -> Integer -> BuiltinByteString

Convert a BuiltinInteger into a BuiltinByteString, as described in CIP-0087. The first argument indicates the endianness of the conversion and the third argument is the integer to be converted, which must be non-negative. The second argument must also be non-negative and it indicates the required width of the output. If the width is zero then the output is the smallest bytestring which can contain the converted input (and in this case, the integer 0 encodes to the empty bytestring). If the width is nonzero then the output bytestring will be padded to the required width with 0x00 bytes (on the left for big-endian conversions and on the right for little-endian conversions); if the input integer is too big to fit into a bytestring of the specified width then the conversion will fail. Conversion will also fail if the specified width is greater than 8192 or the input integer is too big to fit into a bytestring of length 8192.

Conversions

fromBuiltin :: FromBuiltin arep a => arep -> a

toBuiltin :: ToBuiltin a arep => a -> arep