PlutusTx.These

Documentation

data These a b

A These a b is either an a, or a b or an a and a b. Plutus version of These.

Constructors

This a

That b

These a b

type Rep (These a b)

Instance details

Defined in PlutusTx.These

type Rep (These a b) = D1 ('MetaData "These" "PlutusTx.These" "plutus-tx-1.60.0.0-inplace" 'False) (C1 ('MetaCons "This" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 a)) :+: (C1 ('MetaCons "That" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 b)) :+: C1 ('MetaCons "These" 'PrefixI 'False) (S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 a) :*: S1 ('MetaSel ('Nothing :: Maybe Symbol) 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 b))))

type Unroll (These a b)

Instance details

Defined in PlutusTx.These

type Unroll (These a b) = Insert (These a b) (GUnroll (IfStuckRep (RepIsStuckError (These a b) :: Type -> Type) (Rep (These a b))))

these :: (a -> c) -> (b -> c) -> (a -> b -> c) -> These a b -> c

theseWithDefault :: a -> b -> (a -> b -> c) -> These a b -> c

Consume a 'These a b' value.